Rework linking offsets to respect float-alignment.
Compiled code blocks look like

-1. manifest-vector
 0. manifest-nm-vector
 1. code
 ...
 n. object
 ...

Here offset -1 is the address in the tagged object, and offset 0 is
the offset obtained by system-vector-ref.  The object is allocated so
that offset 0 is aligned to float-alignment, but the code is
generated under the assumption that the code is aligned.  Change it
so that we move the code forward as much as is needed for
float-alignment.  E.g., if scheme-object-width is 64 and
float-alignment is 128, we get

-1. manifest-vector
 0. manifest-nm-vector
 1. padding
 2. code
 ...
 n. object
 ...

While here, tidy up the logic and sync the cross-compiler.