Teach lap generator to prefer backward branches over forward ones.
Disabled for C and SVM.  Enabled for x86.

Modern x86 and arm CPUs statically predict forward branches to be not
taken, and backward branches to be taken.  SVM predicts branches to
be not-taken, period, and maybe older CPUs are like that too.  For C
we'll leave it to the C compiler to decide.

There are CPUs with explicit branch prediction hints in the
instruction -- perhaps if someone ever makes a powerpc back end,
we'll want to teach the lap generator to pass that on to the branch
generation code.