WIP: Split compiled entries and compiled return addresses.
This way, a compiled entry can be a pointer to a PC offsets, and a
compiled return addresses can be a pointer to a PC, so we save the
trouble of an indirection for returns.

Future enhancement: Make compiled return addresses point to
PC-relative JMP instructions, so that we can use

  ;; (assign (register #x123) (cons-pointer tag (entry:continuation cont)))
  (CALL (@PCR get-pc))
  (JMP (@PCR cont))
get-pc:
  (OR Q (@R ,rsp) (&U ,tag))
  ...
  (JMP (@PCR uuo-link))

for calls, and

  ;; (pop-return)
  (AND Q (@R ,rsp) (R ,regnum:datum-mask))
  (RET)

for returns, with assistance from the return address stack branch
predictor.

(Only on x86-64 for now.)