Add some fixnum tests to exercise the new rules.
Only does +, -, AND, (I)OR, XOR, and NOT for now.  Should be taught
about quotient, remainder, *, and compositions of + and * to exercise
the LEA rules.  But this is already so large the compiler chokes on
it and I had to uncomment it in small sections to run it, so I'm not
wiring it up to the build for now.  Maybe later it can be split up
and wired into the build.