Set port coding to ISO-8859-1 for IMAP.
This is a sleazy workaround: sometimes the input is really encoded as
UTF-8, but passing it through Scheme `strings' as ISO-8859-1 works
out to approximately the not-wrong thing.  Really we should be
operating on octets, not on characters-as-if-they-were-octets.

(cherry picked from commit c4109d9287476d1f432fc411a040d24ce92a8675)