Don't choose the flo:+inf.0/flo:-inf.0 implementations at cold load.
This is wrong because in our public distributions, the cold load is done on
Ubuntu, but the code is supposed to run on any appropriate unix system.

The fix is to make the selection when the band is restored, guaranteeing that
it's running on the target system.

Manual cherry-pick of 9b94290f82a99dfdb63ae94defcc8a92f2c0514d.