Implement mechanism for hidden global variables.
This is a kludge but is the simplest way to achieve what I need.  This allows
macros to rely on "unexported" variables that aren't inadvertently captured
during expansion.  A number of existing macros suffer from this problem,
particularly in sealed library environments.

For a good example, see macro-deps in (runtime library standard), which
shouldn't be necessary.

Right now each macro has to be hand-crafted to use this mechanism.  It would be
better if it were automatic, which I hope to implement in time.  This could be
achieved if the syntaxer had more detailed information about the definition and
usage environments, which would be useful in any case.

Additionally, doing beta substitution right will require a similar mechanism,
which might as well be this one.

(cherry picked from commit ca1dc586b2b73c66602b7c35d8b2f54c2bbf548a)